erros corrigidos 33/46

+++ src/pt/ulusofona/aed/deisiworldmeter/Populacao.java
@@ -3,17 +3,16 @@ package pt.ulusofona.aed.deisiworldmeter;
 public class Populacao {
     int id;
     int ano;
-    int populacaoMasculina;
-    int populacaoFeminina;
+    long populacaoMasculina;
+    long populacaoFeminina;
     double densidade;
 
-    public Populacao (int id, int ano, int populacaoMasculina, int populacaoFeminina, double densidade) {
+    public Populacao(int id, int ano, long populacaoMasculina, long populacaoFeminina, double densidade) {
         this.id = id;
         this.ano = ano;
         this.populacaoMasculina = populacaoMasculina;
         this.populacaoFeminina = populacaoFeminina;
         this.densidade = densidade;
-
     }
 
     @Override
